third_party/WebKit/LayoutTests/http/tests/loading/preload-image-src.html

<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
</head>
<body>
<script>
    shouldBeTrue("internals.isPreloaded('resources/base-image1.png');");
    shouldBeTrue("internals.isPreloaded('resources/base-image2.png');");
    shouldBeFalse("internals.isPreloaded('resources/base-image3.png');");
    shouldBeFalse("internals.isPreloaded('data:image/gif;base64,R0lGODlhAQABAIAAAP///////yH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==');");
</script>
<div>This test passes if the image (defined in the src attribute) is preloaded, except for data URI.</div>
<img src="resources/base-image1.png">
<img src="resources/base-image2.png" src="resources/base-image3.png">
<img src="data:image/gif;base64,R0lGODlhAQABAIAAAP///////yH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==">
</body>
</html>
