third_party/blink/web_tests/http/tests/loading/preload-image-src-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded('resources/base-image1.png'); is true
PASS internals.isPreloaded('resources/base-image2.png'); is true
PASS internals.isPreloaded('resources/base-image3.png'); is false
PASS internals.isPreloaded('data:image/gif;base64,R0lGODlhAQABAIAAAP///////yH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=='); is false
PASS successfullyParsed is true

TEST COMPLETE

This test passes if the image (defined in the src attribute) is preloaded, except for data URI.
  
